Make the Enhanced Input behavior when input is flushed consistent with the legacy system.
This makes the behavior of Enhanced Input consistent with the Legacy input system when UPlayerInput::FlushPressedKeys is called (loss of focus to the viewport, or it is otherwise called from gameplay code).

Enhanced Input was incorrectly determing if a key was down on the previous frame if the input was flushed because its RawValue would be reset. We now check if the input is currently being flushed on this frame, and if it is then we can determine if the key was down on the previous frame based on it's "Value" in the key state map. If the "Value" is non-zero, but the "RawValue" has been reset to zero, we know it has been reset from the FlushPressedKeys function.

This behavior is Default ON, but can be toggled off in the Project Settings > Enhanced Input.

#jira UE-165730
#rb jodon.karlik
#preflight 6467858975696d2bd7de8254

[CL 25549173 by ben hoffman in ue5-main branch]