Fixed regressions when undoing a deleted OFPA actor post-save
UnloadPackages had been updated to always clear the transaction buffer by default, which resulted in CleanupAfterSuccessfulDelete always clearing the transaction buffer even when it had previously decided not to. CleanupAfterSuccessfulDelete now sets bResetTransBuffer to false when calling UnloadPackages, as it already handles clearing the transaction buffer when needed.

UWorldPartition keeps a list of PinnedActors (which a newly placed actor would be placed in), but would never remove an actor from this list when its FWorldPartitionActorDesc was deleted. This resulted in a newly placed, saved, then deleted, saved, and then undone, actor being unloaded once it was saved after the undo, as PinnedActors thought it was already tracking the actor so would skip taking a new reference which allowed the actor ref-count to reach zero and be unloaded.

#jira
#rb Dave.Belanger, Patrick.Enfedaque

[CL 25592379 by jamie dale in ue5-main branch]