Landscape : Change weight painting strength
Add a gamma function to the paint tools strength to make more of the slider range useful for painting. Setting: PaintStrengthGamma, defaults to 2.2 (1.0 should replicate old behavior)
Added option to disable the painting influence map which effectlively slows down the initial painting and makes it look like it accelerates. Setting: bDisablePaintingStartupSlowdown, defaults to true (false to return to old behavior)

#jira UE-183299
#rb jonathan.bard
#preflight 646feaa5407983b998343fdd

[CL 25651368 by don boogert in ue5-main branch]