Sequencer: Bitfield boolean properties are now able to use fast property offsets
Bitfields were previously not supported because there was no exposed way of determining how an FBoolProperty is mapped in memory. Using fast pointer offsets would have previously caused us to always stomp a full byte over the address. Since we now support meta-data on property components (which is already used by colors and vectors), we can use the FBoolProperty when resolving the property to detect the bitoffset if FBoolProperty::IsNativeBool is false. (see usage of CountTrailingZeros64 inside FBoolHandler)

For my test this pulls down the cost of assigning boolean flags for PostProcessingOverrides from 15us to 1.7us.

Other miscellaneous changes included:
  - Added a way for traits to tell the component handlers that they are not composite property types. This causes them to use direct property setters rather than requiring a composite temporary, and prevents the handler from instantiating partial property setter code.
  - Moved custom accessor functions into their own file. Added one for ULightComponent::VolumetricScatteringIntensity.

#rb Ludovic.Chabant, Max.Chen
#preflight 647907a6e319748a835fbd06

[CL 25760044 by andrew rodham in ue5-main branch]