XRScribe: proper extension filtering in capture layer
Previously, we'd just passs all extensions in/out thru the capture layer to the runtime. XRScribe does not support the full set of OpenXR extensions yet, so the capture layer has to gate what it exposes to the engine.

Once we receive the loader GetProcAddr, we use it to fetch all instance extensions, and create a subset against the capture layer supported extensions. That subset is then exposed to the engine.

#jira UE-186917
#rb Jeff.Fisher, Christopher.Fiala
#rnx

[CL 25827486 by robert srinivasiah in ue5-main branch]