UE-187439 - Variables dragged and dropped to the node with split struct pins are not added
CL 12611535 disabled drag and drop on nodes and their pins to prevent a reconstruction crash. That crash was actually due to drag and dropping a variable onto a *node* that had split pins. By doing that, the variable reference changed, which tripped up node reconstruction since it discovers an inconsistent pin count. However, it's actually safe to drag and drop a variable to a pin connection of a split node. This doesn't affect the underlying variable reference, or introduce any other behavior that might confuse node reconstruction.

The fix here simply removes the additional split pin criteria from DroppedOnPin.

#jira UE-187439
#rb phillip.kavan

[CL 25829105 by dave jones2 in ue5-main branch]