EpicGames.Core: Set a large max degree of parallelism in Json tracer tests.
The default value here is -1, which is equivalent to Environment.ProcessorCount when calling Parallel.ForEachAsync(). This results in a large number of async tasks calling Task.Delay() in sequence, rather than actually running asynchronously, causing the test to take > 60 seconds on machines with low core counts (such as a VM). Increasing this setting to eliminate sequencing allows allows test to complete in < 50ms.

#fyi bryan.sefcik

[CL 25844877 by Ben Marsh in ue5-main branch]