Chaos - ImplicitObjectUnion BVH collision detection optimizations
We now have two types of particle-pair midphase. One which pre-builds a list of all potentially overlapping shape pairs (this already existed), and a new more general purpose one. The general purpose one does not cache any information or prefilter any pairs, but it will use the BVH in the collision hierarchy if there is one.

ImplicitObjectUnion now has its own implementation of a BVH (currently a loose binary tree) to avoid having to make major changes to TBoundingVolumeHierarchy which did not fit its needs.

Midphases are no longer held in an ObjectPool although this could be made to work again (except we now need a pool per midphase type)


#rb none

[CL 25870714 by chris caulfield in ue5-main branch]