[Remote Control] Remote Control Logic - Adds Object controllers and Custom Texture controller
Adds support for some Object controllers:
- Static Mesh
- Material Interface
- Texture (as a custom controller)

Adds support for "custom" controllers, which can be considered as a specialized version of existing controller types
- Custom Texture controller is a specialized string controller loading a texture from a file path or from an asset, through a dedicated widget.

Known issues: transaction are not handled properly. It was difficult to differentiate between what's the expected RC handling of controllers and exposed properties, and the behavior resulting from undoing actions for these new controllers.

Further development of the Texture controller might include the creation of a dedicated struct e.g. FTextureSelector

#jira UE-184894
#rb Jeremie.Roy

[CL 25874085 by Dario Mazzanti in ue5-main branch]