MoviePipeline: Fixed a crash where a Spawnable camera that had the spawn track partially keyed to not visible would crash when adding the job to the queue. Fixed an issue where camera names were not resolving correctly due to multi-camera support.
The fix is to only set the MainCameraIndex if we actually choose to add the camera, as otherwise we were running into situations where the MainCameraIndex would be set, but the camera would not be added, and it would result in an Out of Bounds array access.

Fixed an issue where camera names tokens did not always resolve correctly when using non multi-camera support. This was caused by the placeholder "main camera" entry being pushed into the sidecar array, and then some parts of the rendering were looking up data from the sidecar array (when only trying to render a single camera). Instead, when using single-camera rendering mode we should always use the existing Inner Name built into the UMoviePipelineExecutorShot, and runtime lookups should use the player camera (as Sequencer is animating onto this). We should only look at the sidecar data when rendering all cameras.


#jira UE-185716

[CL 25881669 by Matt Hoffman in ue5-main branch]