Cameras: change the default aspect ratio constraint to maintain Y FOV
Since camera FOV is now always handled as horizontal FOV, changing this setting isn't as disruptive as it used to be, and this now sets a sane default in an age of widescreen monitors. The only thing that might break is any custom gameplay code that is trying to compensate for horizontal FOVs prior to the first change. To make sure nothing breaks in FN we specify the old aspect ratio axis constraint there.

Github pull request 9004

#jira UE-146969
#rb max.chen

[CL 25996500 by ludovic chabant in ue5-main branch]