Fixed a string lifetime bug.
When trying to find the longest matching string for column names in TEDS it would occassionaly just return "column" even if a better match existed. This was because the column label that was retrieved was incorrectly stored so went out of scope and as a result the string retrieved from it was also no longer available. Due to timing of memory reuse this issue wouldn't always show up. The text object is now correctly stored, fixing the the invalid string reference retrieved from it.

#jira UE-186695
#rb han.chu jamie.dale

[CL 26052911 by ronald koppers in 5.3 branch]