Non-immediate command lists always record commands, even if bypass mode is requested.
This is because currently QueueAsyncCommandListSubmit() only supports command lists with their own local contexts for parallel translation. When parallel translation is not used, these commands need to be replayed on the immediate context, but bypass mode would create a local context and translate commands into it, and this context is never finalized and submitted.

Also removed PLATFORM_RHITHREAD_DEFAULT_BYPASS, since bypass should be off by default on all platforms. Using this assumption, the Bypass() methods return compile-time constants in test and shipping builds, so all the bypass code can be compiled out.

#jira UE-185658
#rnx
#rb Luke.Thatcher

[CL 26058493 by mihnea balta in 5.3 branch]