Optimize UClass::FindFunctionByName by extending the function cache (prevously named SuperFuncMap) to include functions in this class in addition to superclasses/interfaces. This means we do a single lock + map lookup instead of needing two if the function isn't found in this class.
This saves 100us on the gamethread, and also helps on async loading to some extent. This is a memory tradeoff since the cache is larger. It adds around ~170KB on average, increasing mem usage of this function from 1.82MB to 1.99MB (in a replay benchmark, as measured with a custom LLM tag).

#tests A/B perf tested in a cooked build on console, also tested with validation to ensure results were the same, tested editor, including a full cook.
#rb rob.millar,dan.oconnor

[CL 26063738 by ben woodhouse in 5.3 branch]