Sequencer: register/unregister the camera animation stand-in object from its own module
This is a speculative fix for a crash that happens on shutdown, when it looks like the camera stand-in UClass does not exist anymore. This is possibly due to how we were unregistering it from the sequencer ECS custom accessors from a different module, and so maybe that module was getting unloaded after the one in which the camera stand-in lives. Either way, keeping things in the same module is probably better anyway.

#rb max.chen

[CL 26075265 by ludovic chabant in 5.3 branch]