Sequencer: fix recycling of view-models when track layouts are rebuilt
Recycling should be done once before the track (or track row) layout is built. This is because when track (or track rows) have multiple sections, we want to make the pre-rebuild view-models available for recycling, but those already re-added for the first section shouldn't be recycled when we start the second section. So we do all this first-time recycling outside of the track layout builder, and effectively revert the (wrong) fix from 25693978 which removed the "don't recycle if recycling already exists" logic.

Other changes here include:
- Removing an unused/unimplemented constructor for FHierarchicalModelListRefresher
- Renaming RecycleChildren to ConditionalRecycleChildren to make it clearer that it's not meant to always recycle (it should only recycle the first time a model is encountered)
- Add a call to Pop() after a top level channel has been added. It never really mattered because generally when something has a top level channel, it doesn't have anything below it

#jira UE-188229
#rb max.chen

[CL 26075298 by ludovic chabant in 5.3 branch]