Introduced the ability to bind to an already created query.
This change introduces the ability to bind a function to an already existing query. It takes the provided functions and creates a wrapper around it that can be used by standard RunQuery calls in TEDS. To support the various different query contexts, the query binder now also no longer requires IQueryContext for it's binding and can use any object that has the minimally required functions. For convenience the current root functions still explicity set the appropriate context as this will produce more readable error messages.

Additional notes:
- Several parts of the query binding were simplified by using C++20 concepts.
- Some cleanup to be able to remove the "using" redirections introduced at a temporary fix with the previous code reorganization.

#jira UE-185241
#rb logan.buchy

[CL 26100582 by ronald koppers in 5.3 branch]