MoviePipeline: Optimized the overlapped accumulator for RGBA images. Improvements mostly center around going wider on high-core count machines with potential improvement going from 70ms->5.7ms per accumulation sample (at 4k). The MRQ pipeline stalls from submitting new renders when no accumulators are available (due to being busy accumulating samples in linear fashion). This is generally noticeable as a few frames will render and then MRQ appears to stall while the accumulators finish and catch up, then a few more frames are submitted, then stall, etc.
Timings for some of the various improvements:

ZeroPlanes (used to reset an accumulator after each output frame):
7.8ms -> 3ms (ParallelFor to loop over RGBA(+Weight) planes all at the same time)
3ms -> 1ms (Memset instead of a for loop setting the values pixel by pixel)

PlaneAccumulation (Unpack FP16 value and add results to existing image plane)
15.27ms -> 6.57ms (ParallelFor to loop over RGBA(+Weight) planes all at the same time)
4.48ms -> 0.90ms (Accumulate by Scanline within each plane accumulation)

Fetch (Get the now accumulated data back out of the Accumulator, re-packing it into RGBA)
38ms->7ms (ParallelFor)

AccumulatePixelData
25ms -> 16.1ms
This was the trickiest improvement as the real cost in this function was hiding in other functions. The accumulation function was allocating a 32mb array each frame, but the allocation only appeared to take 315us so it didn't seem high priority compared to the rest of the function taking 9.8ms (9800us). What was not obvious though is that the OS returns the allocation before actually physically allocating the memory and instead it can wait until you access the memory to actually allocate (and potentially clear the allocated memory for security purposes) which means the cost was being hidden in another function (the cost was about 5ms for the allocation). And then when freeing this temporarily allocated 32mb buffer, that takes another 2ms, but that was attributed to the parent function call as it was only happening during the destructor being run. So in reality about 7 of the 9.8ms was the per-frame allocation (which was moved to a persistent member-level allocation).

There remains a few areas for improvement that were not pursued due to the scope of changes they would require. Currently the Surface Readback code creates a copy which costs us about 3ms, and eventually we have to delete this copy (which costs us another 2ms), but both of these costs are on the Task Thread, and because our overall accumulation time has lowered significantly it was deemed not worth pursuing due to the invasive changes required. Another place we could consider optimizing is the extra "WeightPlaneValue". It's always filled with 1.0 (for all weight planes, every time) and is structured this way so we can re-use the RGBA accumulation code. Ideally we wouldn't allocate a spare memory block just to fill it with 1's but the code is too intrusive to change, so we pay the 8mb RAM cost (1080p, per accumulator, so ~50mb total) for now.

#jira UE-180344
#rb Chad.Williams

[CL 26107093 by matt hoffman in 5.3 branch]